+++ biopandas/testutils/__init__.py
@@ -3,3 +3,5 @@
 # License: BSD 3 clause
 # Project Website: http://rasbt.github.io/biopandas/
 # Code Repository: https://github.com/rasbt/biopandas
+
+from .testutils import assert_raises